feat: introduce Artifact model
This commit introduces the `Artifact` model and the
corresponding factory and establishes the relation to
the `Location` model.

What's an *Artifact* you ask?

An `Artifact` is what we used to call an `Object`.
`Object` is a reserved word in PHP and can't be used.
I think `Artifact` is a good substitute in our context.