chore: Silence Sass deprecation warnings
As you might have noticed, the dev server spits out
a lot of deprecation warnings related to Sass features.
This is nothing to worry about but it's noisy and
clutters the log.
The team behind the Dart Sass implemention is aware
of this and is working on an update to fix this, as
is the Bootstrap team.

This commit updates the Vite config to silence these
warnings.

- https://github.com/twbs/bootstrap/issues/40962
- https://getbootstrap.com/docs/5.3/getting-started/vite/#configure-vite