feat: Introduce ViewServiceProvider for global template variables
This service provider is responsible for making global variables available
to all Blade view templates throughout the application. It uses Laravel's
`View::share()` method to inject variables that can be accessed directly
in any Blade template without needing to pass them from controllers.

This is bit of an experiment and several other options have been considered
like using a custom configuration in combination with `ENV` variables.

Although the implementation might change in the future, it get's the job
done for now. The provider serves as central place for (more or less)
static values used in views that should not be hard-coded.