Update advanced_search.blade.php
I have added a rudimentary form with two search fields. The upper field allows users to search for device names, while the lower field is used to search for project names. Both fields are labeled with attributes that should be considered when processing the request.

I also removed the use of the '$pageTitle' variable in the first <h1> tag because the value didn’t seem appropriate. Instead, I simply used “Advanced Search”.