Update DeviceSeeder.php
Removed attributes that are no longer needed since they are now included in the CSV files. This simplifies the generic usage of the seeder. The IDs for the attributes 'last_edit_by' and 'institution_id' are intended to remain fixed in the future.

Also, the description of the private method has been adjusted.