feat(ui): make app logo parts stylable in nested selectors
CSS had trouble reaching `paths` elements in SVGs that are
referenced via `use`. This made it impossible to override the
individual colors of the logo parts in the navigation bar (or
any other nested selctor).

The coloring for the logo was refactored using CSS custom
properties to overcome this obstacle.

The default behavior is that all colors in the logo are the same
as that of the parent element.

- `--app-logo-color`: overrides all colors
- `--app-logo-primary-color`: overrides the ring color
- `--app-logo-accent-color`: overrides the spark color

Example from the navigation bar:

```css
.navbar .app-logo {
    // Override just the ring color so that it doesn't changes
    // on hover. The spark still inhetits it's color from the
    // parent link element and changes color on hover.
    --app-logo-primary-color: var(--bs-navbar-color);
}
```

This was not possible before. A selector like `.navbar .app-logo-accent`
doesn't match the `path` in the SVG for whatever reason.