fix: avoid empty class attribute in link component
This is more of a refactor (no real bug) but it felt so wrong to just
let this one pass.
Laravels `class` method produces an empty attribute if the attribute
is not passed to the component and none of the conditional classes apply.
There's a build-in `filter` method for attributes but it's not documented
well and I wanted to avoid filtering out any other optional attributes.
Seems to be a lot of code just to compile the classes attribute but it now
works as intended.