feat: add more icons to the sprite
Added the following icons from icons.getbootstrap.com to the SVG sprite
to have some flexibilty.

- Columns Icon
- Controller Icon
- Crosshair Icon
- Database Add Icon
- Gear Wide Connected Icon
- Rocket Takeoff Icon
- Stack Icon
- Three Dots Icon
- Three Dots Vertical Icon
- Vignette Icon
- Vinyl Icon
- Window Sidebar Icon

Unused icons will be removed as soons as we have decided which ones to use.
Some of the icons a general purpose (like "Three Dots" for menu toggles)
and others are meant to represent entities like laser devices.