Replace inputform_artifacts view with inputform_artifact
Removed the inputform_artifacts.blade.php view (copied the inputform_material.blade.php in it for a rough structure) and added a new inputform_artifact.blade.php view with a simplified form for adding a new artifact. The new form omits material selection and focuses solely on artifact name input.