Fixed the form in the view so that it can save data.
In the controller, I added the same names as in the view. In the web.php file, I also added Route::post so that the data can be added to the database. In 'inputform_project.blade.php', I edited the form caption to use the method 'POST' to connect the data saving to the route.