correct cooling filter logic and minor improvements
I removed unnecessary function "key" from line 21. I also fixed a comparison mistake in line 63 (changed '<=' to '>='). I also completed the implementation of the cooling filter: it now distinguishes between devices with and without cooling. I decided to put the devices with no cooling information in the same category as devices with no cooler. To-Do: This decision still needs to be validated by the stakeholders.