chore: add wiki to .gitignore
GitHub Wikis live in their own Git repository usually named
`<user>/<repo>.wiki.git`. This seperation can make edits on the wiki
a bit cumbersome.

`/wiki` was added to `.gitignore` so that the wiki can be cloned into
a subdirectory of the main project without beeing tracked as part of the
code base.